// frontend/src/openapi/models/EvalRunOperationalStat.ts
export interface EvalRunOperationalStat {
    /**
     *
     * @type {string}
     * @memberof EvalRunOperationalStat
     */
    variantId: string;
    /**
     *
     * @type {string}
     * @memberof EvalRunOperationalStat
     */
    label: string;
    /**
     *
     * @type {number}
     * @memberof EvalRunOperationalStat
     */
    sampleCount: number;
    /**
     *
     * @type {number}
     * @memberof EvalRunOperationalStat
     */
    genCost: number;
    /**
     *
     * @type {number}
     * @memberof EvalRunOperationalStat
     */
    evalCost: number;
    /**
     *
     * @type {number}
     * @memberof EvalRunOperationalStat
     */
    totalCost: number;
    /**
     *
     * @type {number}
     * @memberof EvalRunOperationalStat
     */
    genLatencyMs: number | null;
    /**
     *
     * @type {number}
     * @memberof EvalRunOperationalStat
     */
    evalLatencyMs: number | null;
    /**
     *
     * @type {number}
     * @memberof EvalRunOperationalStat
     */
    totalTokens: number | null;
}